+++ packages/extension-drag-handle-react/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @tiptap/extension-drag-handle-react
 
+## 3.16.0
+
+### Patch Changes
+
+- Updated dependencies [9516ee0]
+  - @tiptap/react@3.16.0
+  - @tiptap/extension-drag-handle@3.16.0
+  - @tiptap/pm@3.16.0
+
+## 3.15.3
+
+### Patch Changes
+
+- @tiptap/extension-drag-handle@3.15.3
+- @tiptap/react@3.15.3
+- @tiptap/pm@3.15.3
+
+## 3.15.2
+
+### Patch Changes
+
+- Updated dependencies [d8ca9bf]
+  - @tiptap/react@3.15.2
+  - @tiptap/extension-drag-handle@3.15.2
+  - @tiptap/pm@3.15.2
+
+## 3.15.1
+
+### Patch Changes
+
+- @tiptap/extension-drag-handle@3.15.1
+- @tiptap/pm@3.15.1
+- @tiptap/react@3.15.1
+
+## 3.15.0
+
+### Patch Changes
+
+- @tiptap/extension-drag-handle@3.15.0
+- @tiptap/react@3.15.0
+- @tiptap/pm@3.15.0
+
 ## 3.14.0
 
 ### Patch Changes